[Task-11] Update test
